Move UTF8 funcs into new modules/charset package
